# Conflicts:
#	.mcreator/workspaceBackups/caerula_arbor.mcreator-backup_20250512_224232
#	src/main/java/com/apocalypse/caerulaarbor/block/AnchorMediumBlock.java
#	src/main/java/com/apocalypse/caerulaarbor/entity/BoneFishEntity.java
#	src/main/java/com/apocalypse/caerulaarbor/entity/RunFishEntity.java
#	src/main/java/com/apocalypse/caerulaarbor/event/PlayerEventHandler.java
#	src/main/java/com/apocalypse/caerulaarbor/init/ModItems.java
#	src/main/java/com/apocalypse/caerulaarbor/init/ModTabs.java
#	src/main/java/com/apocalypse/caerulaarbor/item/RescissionItem.java
#	src/main/java/com/apocalypse/caerulaarbor/item/relic/cursed/CaerulaAnimusItem.java
#	src/main/java/com/apocalypse/caerulaarbor/item/relic/epic/HandOfBrandingItem.java
#	src/main/java/com/apocalypse/caerulaarbor/item/relic/epic/KingsCrownItem.java
#	src/main/java/com/apocalypse/caerulaarbor/item/relic/normal/AntiquatedSheetMusicItem.java
#	src/main/java/com/apocalypse/caerulaarbor/item/relic/normal/StoneGargoyleItem.java
#	src/main/java/com/apocalypse/caerulaarbor/procedures/DeductPlayerSanityProcedure.java
#	src/main/java/com/apocalypse/caerulaarbor/procedures/DetectActivityProcedure.java
#	src/main/java/com/apocalypse/caerulaarbor/procedures/PlayergetupFuncProcedure.java
#	src/main/resources/assets/caerula_arbor/lang/en_us.json
#	src/main/resources/assets/caerula_arbor/lang/zh_cn.json

+++ src/main/java/com/apocalypse/caerulaarbor/procedures/GiveLessArmorProcedure.java
@@ -13,7 +13,7 @@ public class GiveLessArmorProcedure {
 			if ((obj instanceof LivingEntity _livEnt && _livEnt.hasEffect(ModMobEffects.LESS_ARMOR.get()) ? _livEnt.getEffect(ModMobEffects.LESS_ARMOR.get()).getAmplifier() : 0) < limit) {
 				if (obj instanceof LivingEntity _entity && !_entity.level().isClientSide())
 					_entity.addEffect(new MobEffectInstance(ModMobEffects.LESS_ARMOR.get(), 300,
-							(int) ((obj instanceof LivingEntity _livEnt && _livEnt.hasEffect(ModMobEffects.LESS_ARMOR.get()) ? _livEnt.getEffect(ModMobEffects.LESS_ARMOR.get()).getAmplifier() : 0) + 1), false, true));
+							(obj instanceof LivingEntity _livEnt && _livEnt.hasEffect(ModMobEffects.LESS_ARMOR.get()) ? _livEnt.getEffect(ModMobEffects.LESS_ARMOR.get()).getAmplifier() : 0) + 1, false, true));
 			}
 		} else {
 			if (obj instanceof LivingEntity _entity && !_entity.level().isClientSide())
